chore: add kit helper
